fix(customer-portal): centralize upstream-error body extraction, fix Vary header
Every upstream client (registry, updates, scim, productconsumption,
aichatagent, usermanagement) built its own non-2xx apierror.Error by falling
back to a raw response excerpt when the body wasn't {"message": "..."}
shaped — the same raw-content-leak risk already fixed for the entity client.
Add apierror.NewUpstreamError as the one shared constructor (Body is the
parsed message, or empty) and route every client through it instead of
duplicating the logic. Also switch cors.go's Vary header to Add so it
doesn't clobber a Vary value a downstream handler already set.
