+++ apps/customer-portal/backend-v2/internal/usermanagement/usermanagement.go
@@ -136,31 +136,8 @@ func (c *Client) ValidateProjectContact(ctx context.Context, req ValidationPaylo
 	case http.StatusAccepted:
 		return nil, false, nil
 	case http.StatusConflict:
-		return nil, true, extractErrorMessage(statusCode, raw)
+		return nil, true, apierror.NewUpstreamError(statusCode, raw)
 	default:
-		return nil, false, extractErrorMessage(statusCode, raw)
+		return nil, false, apierror.NewUpstreamError(statusCode, raw)
 	}
-}
-
-// upstreamErrorBody is the {"message": "..."} shape every non-success
-// response from this service carries.
-type upstreamErrorBody struct {
-	Message string `json:"message"`
-}
-
-// extractErrorMessage builds an *apierror.Error whose Body is the upstream
-// service's own "message" field when present, so callers can surface the
-// specific reason (e.g. "Contact already exists") rather than a generic
-// fallback.
-func extractErrorMessage(statusCode int, raw []byte) error {
-	var body upstreamErrorBody
-	if err := json.Unmarshal(raw, &body); err == nil && body.Message != "" {
-		return &apierror.Error{StatusCode: statusCode, Body: body.Message}
-	}
-	const maxErrBody = 512
-	excerpt := raw
-	if len(excerpt) > maxErrBody {
-		excerpt = excerpt[:maxErrBody]
-	}
-	return &apierror.Error{StatusCode: statusCode, Body: string(excerpt)}
 }
